Clean up the spec to use externally maintained macros and correct deps
Current versions of debbuild can now consume complex externally maintained
macros, so we're going to use those from now on. This further reduces
the differences between RPM and Debian distributions in terms of amount
of boilerplate.

Also, clarify some dependencies at runtime so that we get the right
things everywhere.