Encapsulate "bare words" in expressions for RPM 4.16+ compatibility
In RPM 4.16, the spec parser was changed to no longer accept expressions
with so-called "bare words". That is, expressions like the following are
no longer accepted:

%if %{_vendor} == "debbuild"

This triggers the following error:

error: bare words are no longer supported, please use "...": redhat == "debbuild"

Thus, all the conditionals where this was happening are now quoted
on the left hand term, just like on the right hand term.

This is fully backwards compatible to older RPM versions and debbuild,
so there's no incompatibility from older distributions.